+++ src/main.rs
@@ -14,7 +14,6 @@ mod runner;
 mod snapshot;
 
 use call_stack::StackKind;
-use cirru_edn;
 use dirs::home_dir;
 use primes::Calcit;
 use std::fs;
@@ -87,7 +86,7 @@ fn main() -> Result<(), String> {
         match result {
           Ok(v) => {
             let duration = Instant::now().duration_since(started_time);
-            println!("took {}ms: {}", duration.as_micros() as f32 / 1000.0, v);
+            println!("took {}ms: {}", duration.as_micros() as f64 / 1000.0, v);
             Ok(())
           }
           Err(failure) => {
